Fix nesting when making OpenAPI arrays and objects optional
Prior to this change, the mangling of optional arrays and objects
misses "related" properties. For example, given this OAS:

  {
    "name": "someProperty",
    "required": false,
    "schema": {
      "type": "array",
      "items": {
        "type": "string"
      }
    }
  }

The resulting MCP tool schema would be:

  {
    "someProperty": {
      "items": {
        "type": "string"
      },
      "anyOf": [
        {
          "type": "array"
        },
        {
          "type": "null"
        }
      ]
    }
  }

This is invalid (and e.g. OpenAI fails on it). Instead, the
"items" property should be a part of the now-nested "array" type:

  {
    "someProperty": {
      "anyOf": [
        {
          "type": "array",
          "items": {
            "type": "string"
          },
        },
        {
          "type": "null"
        }
      ]
    }
  }

This commit looks at all the properties defined in the JSON schema
to belong to either array or object type and nests them.