fix(azure): Example patch for logic required for Azure authentication with the OAuthProxy.
This is obviously a hack, but it's been created to prove that
authentication with Azure has to be rewritten, due to Azure quirks.

In short it achieves the following:
- When requesting a token from Azure, the scope must be defined with the
  client ID in it.
  - Example: `api://{client_id}/user_impersonation` as a scope will construct this URL for Azure: https://login.microsoftonline.com/{tenant_id}/oauth2/v2.0/authorize?response_type=code&client_id={client_id}&redirect_uri=http://localhost:8000/auth/callback&state={state}&scope=api://{client_id}/user_impersonation
- The token is exchanged for a token
- When validating the token, the token only has `user_impersonation`,
  but the audience is set to `client_id`.
  - Example: {"aud": "<client id>", "scp": "user_impersonation", ...}

So we must decouple required scope for the redirect URI to Azure, and
the scopes verified in the token validation + scope check for endpoints.

This patch "fixes" the problem with simple if-statements and some scope
stripping, but is not a
long term solution. How ever, it allows me to have a workshop with
working Azure auth and OBO flow on Monday. :)

The relevant issue for this commit can be found here: https://github.com/jlowin/fastmcp/issues/1663#issuecomment-3235993369