Expose valid scopes from OAuthProxy metadata
OAuthProxy presently creates a ClientRegistrationOptions instance with `enabled=True`, but does not pass through and `valid_scopes`. Consequently, the metadata for OAuth well-known endpoints does not include valid scopes.

There are some clients, like Claude, which looks at the advertised scopes and use this to set the list of scopes that will be requested for client registration and during authorization. If you want Claude (and possibly others) to request scopes, then these need to be set.

This PR allows passing in a list of valid scopes when you create the OAuthProxy instance and attaches these to the ClientRegistrationOptions instance that is created. This is all that is required for the metadata to be exposed in the well-known routes.

I thought about setting `valid_scopes` equal to the token verifier's `required_scopes`, but I can think of scenarios where you might want to advertise a larger set of scopes than what is required. For example, if you have read & write as valid scopes. You might only require read scope. The MCP server might then only expose a subset of tools to a user with read-only scope. The way this PR is set up, `valid_scopes` and `required_scopes` are separate so it would accommodate that situation.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>