Protected resource advertises valid scopes, if available
With #1717, the `.well-known/oauth-authorization-server` was enabled to list a set of supported scopes, not just the scopes that are required (required scopes can be a "minimal" set, while supported scopes may be a larger set as some may be optional).

This PR follows up on #1717 and enables the same support for `.well-known/oauth-protected-resource`. When creating this resource route, we pulls the valid_scopes from the ClientRegistrationOptions object, if it's set. If it's not set, we fall back to required_scopes, which was the previous behavior.

This keeps the supported scopes list in sync across the two endpoints.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>